+++ packages/compiler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @generaltranslation/compiler
 
+## 1.3.15
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+
 ## 1.3.14
 
 ### Patch Changes