+++ packages/next/CHANGELOG.md
@@ -1,5 +1,16 @@
 # gt-next
 
+## 6.16.18
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+  - gt-i18n@0.8.8
+  - @generaltranslation/compiler@1.3.15
+  - gt-react@10.19.7
+  - @generaltranslation/supported-locales@2.0.67
+
 ## 6.16.17
 
 ### Patch Changes