+++ packages/react-native/CHANGELOG.md
@@ -1,5 +1,14 @@
 # gt-react-native
 
+## 10.19.7
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+  - @generaltranslation/react-core@1.8.9
+  - @generaltranslation/supported-locales@2.0.67
+
 ## 10.19.6
 
 ### Patch Changes