+++ packages/supported-locales/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @generaltranslation/supported-locales
 
+## 2.0.67
+
+### Patch Changes
+
+- Updated dependencies [[`ce0933a`](https://github.com/generaltranslation/gt/commit/ce0933ab102d34a0c38634f7c2b0d634c9a620a8)]:
+  - generaltranslation@8.2.9
+
 ## 2.0.66
 
 ### Patch Changes